srv_reader/requirements.txt
grpcio===1.63.0
grpcio-tools===1.63.0